+++ src/components/shoppingCart/CartItemList/index.tsx
@@ -1,7 +1,7 @@
-import { useRecoilState } from "recoil";
+import { useRecoilState, useRecoilValue } from "recoil";
 import { isAllCartItemSelectedState } from "@/stores/cartItemSelections";
 
-import useCartItems from "@/hooks/useCartItems";
+import useCartItems from "@/hooks/carts/useCartItems";
 
 import CartItemCard from "@/components/shoppingCart/CartItemCard";
 import { CheckButton } from "@/components/button";
